+++ src/Render/NeoFormRenderer.php
@@ -3,9 +3,6 @@
 namespace Efabrica\NeoForms\Render;
 
 use Efabrica\NeoForms\Build\NeoForm;
-use Efabrica\NeoForms\Control\ControlGroupBuilder;
-use Efabrica\NeoForms\Control\Div;
-use Efabrica\NeoForms\Render\NeoInputRenderer;
 use Latte\Engine;
 use Nette\Forms\ControlGroup;
 use Nette\Forms\Controls\BaseControl;