+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - NeoFormFactory
 - AbstractForm
 - more documentation
+- merge AbstractFormFactory into AbstractForm
 
 ## [0.0.2] - 2022-07-01
 ### Fixed