+++ CHANGELOG.md
@@ -5,15 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+
+## [1.0.0]
 ### Added
 - Nette Form Renderer (you can use `{control form}` now)
 - NeoFormFactory
-- Div, Row, Col
 - MultiSelectBox
 - Changed namespaces and directory structure
 - readonly renderer
 - NeoForm class
-- ControlGroupBuilder
+- ControlGroupBuilder (->group(), ->row(), ->col())
+- direct methods instead of container extensions
 - Tags::getTagValues
 - StaticTags -> allowCustomTags
 - {formGroup}
@@ -29,5 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.0.1 - 2022-06-30
 - Initial release
 
-[Unreleased]: https://git.efabrica.sk/libraries/neoforms/compare/0.0.2...master
+[Unreleased]: https://git.efabrica.sk/libraries/neoforms/compare/1.0.0...master
+[1.0.0]: https://git.efabrica.sk/libraries/neoforms/compare/0.0.2...1.0.0
 [0.0.2]: https://git.efabrica.sk/libraries/neoforms/compare/0.0.1...0.0.2